Fix store (saga) bug (losing the id on delete request to API)
I don't know why, but I thought the API is going to return the id of the
deleted entry as a response to delete request. It's not.
Because of this wrong assumption, deletion wasn't occurring in-memory.
Item was deleted from database, because the request itself was correct.
But id was not returning, so redux was able to discover the item was missing
only when fetching all or this specific one.